requirements.txt
Faker
omegaconf
pandas
marshmallow_dataclass==8.4.1
numpy
hydra_core
typeguard
hyperopt
matplotlib
pytest
dataclasses==0.6
scikit_learn==1.0
